# api/contract/openapi.yaml
openapi: 3.0.0
paths:
  /sessions:
    post:
      operationId: SessionController_create
      parameters: []
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Create a new WhatsApp session
      tags:
        - Session
    get:
      operationId: SessionController_getAll
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: List all sessions (from database)
      tags:
        - Session
  /sessions/{sessionId}:
    get:
      operationId: SessionController_getStatus
      parameters:
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get session status
      tags:
        - Session
    delete:
      operationId: SessionController_delete
      parameters:
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Delete a session and remove all data
      tags:
        - Session
  /sessions/{sessionId}/logout:
    post:
      operationId: SessionController_logout
      parameters:
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Logout from WhatsApp and delete session
      tags:
        - Session
  /sessions/{sessionId}/reconnect:
    post:
      operationId: SessionController_reconnect
      parameters:
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Force reconnect a session from stored auth state
      tags:
        - Session
  /sessions/{sessionId}/messages/{jid}:
    get:
      operationId: SessionController_getMessages
      parameters:
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: 'Number of messages (default: 25)'
          schema:
            type: number
        - name: jid
          required: true
          in: path
          description: Chat JID
          schema: {}
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get message history for a chat (from database)
      tags:
        - Session
  /sessions/{sessionId}/contacts:
    get:
      operationId: SessionController_getContacts
      parameters:
        - name: offset
          required: false
          in: query
          description: 'Offset (default: 0)'
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: 'Limit (default: 50)'
          schema:
            type: number
        - name: search
          required: false
          in: query
          description: Search by name or JID
          schema:
            type: string
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get stored contacts (from database)
      tags:
        - Session
  /sessions/{sessionId}/chats:
    get:
      operationId: SessionController_getChats
      parameters:
        - name: offset
          required: false
          in: query
          description: 'Offset (default: 0)'
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: 'Limit (default: 50)'
          schema:
            type: number
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get stored chats (from database)
      tags:
        - Session
  /sessions/{sessionId}/webhooks/logs:
    get:
      operationId: SessionController_getWebhookLogs
      parameters:
        - name: offset
          required: false
          in: query
          description: 'Offset (default: 0)'
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: 'Limit (default: 50)'
          schema:
            type: number
        - name: sessionId
          required: true
          in: path
          description: Session ID
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get webhook delivery logs (from database)
      tags:
        - Session
...
  /{sessionId}/chats:
    get:
      operationId: ChatController_getChats
      parameters:
        - name: offset
          required: false
          in: query
          description: 'Offset (default: 0)'
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: 'Limit (default: 50)'
          schema:
            type: number
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get all chats (from database)
      tags:
        - Chat
...
  /{sessionId}/contacts:
    get:
      operationId: ContactController_getContacts
      parameters:
        - name: offset
          required: false
          in: query
          description: 'Offset (default: 0)'
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: 'Limit (default: 50)'
          schema:
            type: number
        - name: search
          required: false
          in: query
          description: Search by name or JID
          schema:
            type: string
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get all contacts (from database)
      tags:
        - Contact
...
info:
  title: Baileys WhatsApp API
  description: Comprehensive REST API for WhatsApp using Baileys library.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-api-key:
      type: apiKey
      in: header
      name: x-api-key
...